refactor: merged data model with types and renamed VC type
