--- src/4.0/validate/dataModel.ts
@@ -1,154 +1,0 @@
-import z from "zod";
-
-import { ContextUrl } from "../../shared/@types/document";
-
-const baseType = "VerifiableCredential";
-
-// Custom URI validation function
-const URI_REGEX =
-  /^(?=.)(?!https?:\/(?:$|[^/]))(?!https?:\/\/\/)(?!https?:[^/])(?:[a-zA-Z][a-zA-Z\d+-\.]*:(?:(?:\/\/(?:[\w-\.~%\dA-Fa-f!\$&'\(\)\*\+,;=:]*@)?(?:\[(?:(?:(?:[\dA-Fa-f]{1,4}:){6}(?:[\dA-Fa-f]{1,4}:[\dA-Fa-f]{1,4}|(?:(?:0{0,2}\d|0?[1-9]\d|1\d\d|2[0-4]\d|25[0-5])\.){3}(?:0{0,2}\d|0?[1-9]\d|1\d\d|2[0-4]\d|25[0-5]))|::(?:[\dA-Fa-f]{1,4}:){5}(?:[\dA-Fa-f]{1,4}:[\dA-Fa-f]{1,4}|(?:(?:0{0,2}\d|0?[1-9]\d|1\d\d|2[0-4]\d|25[0-5])\.){3}(?:0{0,2}\d|0?[1-9]\d|1\d\d|2[0-4]\d|25[0-5]))|(?:[\dA-Fa-f]{1,4})?::(?:[\dA-Fa-f]{1,4}:){4}(?:[\dA-Fa-f]{1,4}:[\dA-Fa-f]{1,4}|(?:(?:0{0,2}\d|0?[1-9]\d|1\d\d|2[0-4]\d|25[0-5])\.){3}(?:0{0,2}\d|0?[1-9]\d|1\d\d|2[0-4]\d|25[0-5]))|(?:(?:[\dA-Fa-f]{1,4}:){0,1}[\dA-Fa-f]{1,4})?::(?:[\dA-Fa-f]{1,4}:){3}(?:[\dA-Fa-f]{1,4}:[\dA-Fa-f]{1,4}|(?:(?:0{0,2}\d|0?[1-9]\d|1\d\d|2[0-4]\d|25[0-5])\.){3}(?:0{0,2}\d|0?[1-9]\d|1\d\d|2[0-4]\d|25[0-5]))|(?:(?:[\dA-Fa-f]{1,4}:){0,2}[\dA-Fa-f]{1,4})?::(?:[\dA-Fa-f]{1,4}:){2}(?:[\dA-Fa-f]{1,4}:[\dA-Fa-f]{1,4}|(?:(?:0{0,2}\d|0?[1-9]\d|1\d\d|2[0-4]\d|25[0-5])\.){3}(?:0{0,2}\d|0?[1-9]\d|1\d\d|2[0-4]\d|25[0-5]))|(?:(?:[\dA-Fa-f]{1,4}:){0,3}[\dA-Fa-f]{1,4})?::[\dA-Fa-f]{1,4}:(?:[\dA-Fa-f]{1,4}:[\dA-Fa-f]{1,4}|(?:(?:0{0,2}\d|0?[1-9]\d|1\d\d|2[0-4]\d|25[0-5])\.){3}(?:0{0,2}\d|0?[1-9]\d|1\d\d|2[0-4]\d|25[0-5]))|(?:(?:[\dA-Fa-f]{1,4}:){0,4}[\dA-Fa-f]{1,4})?::(?:[\dA-Fa-f]{1,4}:[\dA-Fa-f]{1,4}|(?:(?:0{0,2}\d|0?[1-9]\d|1\d\d|2[0-4]\d|25[0-5])\.){3}(?:0{0,2}\d|0?[1-9]\d|1\d\d|2[0-4]\d|25[0-5]))|(?:(?:[\dA-Fa-f]{1,4}:){0,5}[\dA-Fa-f]{1,4})?::[\dA-Fa-f]{1,4}|(?:(?:[\dA-Fa-f]{1,4}:){0,6}[\dA-Fa-f]{1,4})?::)|v[\dA-Fa-f]+\.[\w-\.~!\$&'\(\)\*\+,;=:]+)\]|(?:(?:0{0,2}\d|0?[1-9]\d|1\d\d|2[0-4]\d|25[0-5])\.){3}(?:0{0,2}\d|0?[1-9]\d|1\d\d|2[0-4]\d|25[0-5])|[\w-\.~%\dA-Fa-f!\$&'\(\)\*\+,;=]{1,255})(?::\d*)?(?:\/[\w-\.~%\dA-Fa-f!\$&'\(\)\*\+,;=:@]*)*)|\/(?:[\w-\.~%\dA-Fa-f!\$&'\(\)\*\+,;=:@]+(?:\/[\w-\.~%\dA-Fa-f!\$&'\(\)\*\+,;=:@]*)*)?|[\w-\.~%\dA-Fa-f!\$&'\(\)\*\+,;=:@]+(?:\/[\w-\.~%\dA-Fa-f!\$&'\(\)\*\+,;=:@]*)*|(?:\/\/\/[\w-\.~%\dA-Fa-f!\$&'\(\)\*\+,;=:@]*(?:\/[\w-\.~%\dA-Fa-f!\$&'\(\)\*\+,;=:@]*)*)))(?:\?[\w-\.~%\dA-Fa-f!\$&'\(\)\*\+,;=:@\/\?]*(?=#|$))?(?:#[\w-\.~%\dA-Fa-f!\$&'\(\)\*\+,;=:@\/\?]*)?$/;
-export const Uri = z.string().regex(URI_REGEX, { message: "Invalid URI" });
-
-export const W3cVerifiableCredential = z.object({
-  "@context": z.union([
-    z.record(z.any()),
-    z.string(),
-    // If array: First item must be baseContext, while remaining items can be string or object
-    z.tuple([z.literal(ContextUrl.v2_vc)]).rest(z.union([z.string(), z.record(z.any())])),
-  ]),
-
-  // [Optional]
-  name: z.string().optional(),
-
-  // [Optional] If string: Must match uri pattern
-  id: Uri.optional(),
-
-  type: z.union([
-    z.string(),
-    // If array: Must have VerifiableCredential, while remaining items can be any string
-    z.array(z.string()).refine((types) => types.includes(baseType), { message: `Type must include ${baseType}` }),
-  ]),
-
-  credentialSchema: z
-    .union([
-      // If object: Must have id match uri pattern and type defined
-      z
-        .object({
-          id: Uri,
-          type: z.string(),
-        })
-        .passthrough(),
-      // If array: Every object must have id match uri pattern and type defined
-      z.array(
-        z
-          .object({
-            id: Uri,
-            type: z.string(),
-          })
-          .passthrough()
-      ),
-    ])
-    .optional(),
-
-  issuer: z.union([
-    // If string: Must match uri pattern
-    Uri,
-    // If object: Must have id match uri pattern
-    z
-      .object({
-        id: Uri,
-      })
-      .passthrough(),
-  ]),
-
-  validFrom: z.string().datetime({ offset: true }).optional(),
-
-  validUntil: z.string().datetime({ offset: true }).optional(),
-
-  credentialSubject: z.union([
-    // If object: Cannot be empty (i.e. minimum 1 key)
-    z.record(z.any()).refine((obj) => Object.keys(obj).length > 0, {
-      message: "Must have at least one key",
-    }),
-    // If array: Every object cannot be empty (i.e. minimum 1 key)
-    z.array(
-      z.record(z.any()).refine((obj) => Object.keys(obj).length > 0, {
-        message: "Must have at least one key",
-      })
-    ),
-  ]),
-
-  credentialStatus: z
-    .object({
-      // If id is present, id must match uri pattern (credentialStatus.id is optional and can be undefined)
-      id: Uri.optional(),
-      // Must have type defined
-      type: z.string(),
-    })
-    .passthrough()
-    .optional(),
-
-  // [Optional] This is at risk of being removed from the w3c spec
-  renderMethod: z.any().optional(),
-
-  termsOfUse: z
-    .union([
-      // If object: Must have type defined. If id is present, id must match uri pattern (termsOfUse.id is optional and can be undefined)
-      z
-        .object({
-          id: Uri.optional(),
-          type: z.string(),
-        })
-        .passthrough(),
-      // If array: Every object must have type defined. If id is present, id must match uri pattern (termsOfUse.id is optional and can be undefined)
-      z.array(
-        z
-          .object({
-            id: Uri.optional(),
-            type: z.string(),
-          })
-          .passthrough()
-      ),
-    ])
-    .optional(),
-
-  evidence: z
-    .union([
-      // If object: Must have type defined. If id is present, id must match uri pattern (evidence.id is optional and can be undefined)
-      z
-        .object({
-          id: Uri.optional(),
-          type: z.string(),
-        })
-        .passthrough(),
-      // If array: Every object must have type defined. If id is present, id must match uri pattern (evidence.id is optional and can be undefined)
-      z.array(
-        z
-          .object({
-            id: Uri.optional(),
-            type: z.string(),
-          })
-          .passthrough()
-      ),
-    ])
-    .optional(),
-
-  proof: z
-    .union([
-      // If object: Must have type defined
-      z
-        .object({
-          type: z.string(),
-        })
-        .passthrough(),
-      // If array: Every object must have type defined
-      z.array(
-        z
-          .object({
-            type: z.string(),
-          })
-          .passthrough()
-      ),
-    ])
-    .optional(),
-});
