+++ src/4.0/validate/index.ts
@@ -1,2 +1,1 @@
 export * from "./context";
-export * from "./dataModel";
